+++ sky-server/src/main/java/com/sky/service/SetmealService.java
@@ -25,7 +25,7 @@ public interface SetmealService {
      * 批量删除套餐
      * @param ids
      */
-    void deleteBatch(List<Integer> ids);
+    void deleteBatch(List<Long> ids);
 
     /**
      * 更新套餐
